Updated to new electron based UI
Quite a few changes, updating public codebase to reflect the more recent GUI.

GUI untested, but feels easier to work with than TKINTER given my lack of knowing python.

The file system as state for the chat program was interesting. It did not however reduce the complexity of implementing new UI's so it was scrapped.